Switched Entity to TS
