+++ application/public/js/game/networking/networkObjects/networkPlayer.js
@@ -1,12 +1,12 @@
-import Entity from "../../objects/entities/entity.js"
+const { Entity } = require("@obesity/components")
 
 
 
 /**
  * Network Player Object
  */
-export default class NetworkPlayer extends Entity {
+export default class NetworkPlayer extends Entity.BaseEntity {
     constructor(scene, x, y, spriteID, name, depth){
-        super(scene, x, y, spriteID, name, depth) // Parent ( Entity ) Setup
+        super(scene, x, y, spriteID, null, name, depth) // Parent ( Entity ) Setup
     }
 }
