+++ components/src/entities/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Game Entities
+ */
+export * from "./entity"
