+++ components/src/index.ts
@@ -2,3 +2,8 @@
  * UI Components
  */
 export * from "./ui"
+
+/**
+ * Entities
+ */
+export * from "./entities"
